Feature/improvements (#6)
* Application Services for Cart and Order, returning DTOs

* Application Services for Cart and Order, returning DTOs

* Moved some logic from Domain Service to Application Service

* updated phpstan level to 9

* improvements

* updated checks.sh

* cleanup

+++ src/ShoppingCart/Application/OrderLineDto.php
@@ -10,7 +10,7 @@ readonly final class OrderLineDto implements JsonSerializable
     public function __construct(
         public string $name,
         public string $sku,
-        public int $unitPrice,
+        public float $unitPrice,
         public int $quantity,
         public float $price
     ) {
